requirements.txt
# Exchange and data
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3

# Technical analysis
pandas-ta==0.3.14b

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Database
SQLAlchemy==2.0.25

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Monitoring and alerts
python-telegram-bot==20.7
schedule==1.2.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-mock==3.12.0
